+++ openapi/__manifest__.py
@@ -6,7 +6,7 @@
     "category": "",
     # "live_test_url": "",
     "images": ['images/openapi-swagger.png'],
-    "version": "13.0.1.1.4",
+    "version": "13.0.1.1.6",
     "application": False,
 
     "author": "IT-Projects LLC, Ivan Yelizariev",