+++ openapi/doc/changelog.rst
@@ -1,6 +1,14 @@
-`1.1.4`
+`1.1.6`
 -------
 - **Fix:** users got same token on installation
+
+`1.1.5`
+-------
+- **Fix:** UTF-8 Decode error, singleton error in check methods
+
+`1.1.4`
+-------
+- **Fix:** no computation In PATCH-method for compute-fields
 
 `1.1.3`
 -------